apps/backend/pyproject.toml
[project]
name = 'agent1-backend'
version = '0.1.0'
description = 'Agent1 FastAPI backend'
readme = 'README.md'
requires-python = '>=3.10'
dependencies = [
  'alembic>=1.13.2',
  'fastapi>=0.115.0',
  'opentelemetry-api>=1.39.1',
  'opentelemetry-instrumentation-fastapi>=0.60b1',
  'opentelemetry-sdk>=1.39.1',
  'psycopg2-binary>=2.9.11',
  'pydantic-settings>=2.5.2',
  'sentry-sdk[fastapi]>=2.53.0',
  'sqlalchemy>=2.0.35',
  'uvicorn>=0.30.6'
]

[project.optional-dependencies]
dev = [
  'mypy>=1.11.2',
  'pytest>=8.3.3',
  'ruff>=0.6.7'
]

[build-system]
requires = ['setuptools>=68', 'wheel']
build-backend = 'setuptools.build_meta'

[tool.setuptools]
package-dir = {'' = 'src'}

[tool.setuptools.packages.find]
where = ['src']

[tool.pytest.ini_options]
addopts = '-q'
testpaths = ['tests']

[tool.ruff]
line-length = 100
target-version = 'py310'

[tool.mypy]
python_version = '3.10'
strict = true
warn_unused_configs = true
